requirements.txt
streamlit
pillow
matplotlib
iapws
pandas
numpy